pyproject.toml
[project]
name = "auth-django"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "psycopg2>=2.9.10",
    "uwsgi>=2.0.29",
]

[dependency-groups]
dev = [
    "coverage>=7.8.0",
    "pre-commit>=4.2.0",
    "pre-commit-hooks>=5.0.0",
    "ruff>=0.11.5",
]

[tool.ruff]
line-length = 120
exclude = [".git", "__pycache__", "venv", ".venv"]
